+++ src/PmeRealSpace.h
@@ -19,30 +19,30 @@ public:
 
   void set_num_atoms(int natoms);
 
-  void fill_charges(double **q_arr, double **q_arr_list, int &q_arr_count,
+  void fill_charges(float **q_arr, float **q_arr_list, int &q_arr_count,
                        int &stray_count, char *f_arr, char *fz_arr, PmeParticle p[]); 
-  void compute_forces(const double * const *q_arr, const PmeParticle p[], 
+  void compute_forces(const float * const *q_arr, const PmeParticle p[], 
                       Vector f[]);
                       
-  void compute_forces_order4_partial(int first, int last, const double * const *q_arr, const PmeParticle p[], 
+  void compute_forces_order4_partial(int first, int last, const float * const *q_arr, const PmeParticle p[], 
                       Vector f[]);                      
 private:
-  void fill_charges_order4(double **q_arr, double **q_arr_list, int &q_arr_count,
+  void fill_charges_order4(float **q_arr, float **q_arr_list, int &q_arr_count,
                        int &stray_count, char *f_arr, char *fz_arr, PmeParticle p[]); 
   template <int order>
-  void fill_charges_order(double **q_arr, double **q_arr_list, int &q_arr_count,
+  void fill_charges_order(float **q_arr, float **q_arr_list, int &q_arr_count,
                        int &stray_count, char *f_arr, char *fz_arr, PmeParticle p[]); 
-  void compute_forces_order4(const double * const *q_arr, const PmeParticle p[], 
+  void compute_forces_order4(const float * const *q_arr, const PmeParticle p[], 
                       Vector f[]);
   template <int order>
-  void compute_forces_order(const double * const *q_arr, const PmeParticle p[], 
+  void compute_forces_order(const float * const *q_arr, const PmeParticle p[], 
                       Vector f[]);
   template <int order> void fill_b_spline(PmeParticle p[]);
 
   const PmeGrid myGrid;
   int N;
-  double *M, *dM;
-  ResizeArray<double> M_alloc, dM_alloc;
+  float *M, *dM;
+  ResizeArray<float> M_alloc, dM_alloc;
 };
